+++ CMakeLists.txt
@@ -6,8 +6,15 @@ set(PROJECT_NAME aig)
 set(PROJECT_DESCRIPTION "Analytical inverse geomery for MPC in bipeds")
 set(PROJECT_URL https://github.com/${PROJECT_NAMESPACE}/${PROJECT_NAME})
 
+# Project options
+option(BUILD_PYTHON_INTERFACE "Build the python binding" ON)
+option(INSTALL_PYTHON_INTERFACE_ONLY "Instal *ONLY* the python bindings" OFF)
+option(SUFFIX_SO_VERSION "Suffix library name with its version" ON)
+
 # Project configuration
-set(PROJECT_USE_CMAKE_EXPORT TRUE)
+if(NOT INSTALL_PYTHON_INTERFACE_ONLY)
+  set(PROJECT_USE_CMAKE_EXPORT TRUE)
+endif()
 set(CUSTOM_HEADER_DIR ${PROJECT_NAME})
 
 # #################################################### Check if the submodule
@@ -31,18 +38,27 @@ endif()
 
 # JRL-cmakemodule setup
 include("${JRL_CMAKE_MODULES}/base.cmake")
+include("${JRL_CMAKE_MODULES}/boost.cmake")
+include("${JRL_CMAKE_MODULES}/python.cmake")
 # ##############################################################################
 
 # Project definition
 compute_project_args(PROJECT_ARGS LANGUAGES CXX)
-project(aig ${PROJECT_ARGS})
-check_minimal_cxx_standard(11 ENFORCE)
+project(${PROJECT_NAME} ${PROJECT_ARGS})
+# check_minimal_cxx_standard(11 ENFORCE)
 
 # Dependencies
 add_project_dependency(pinocchio REQUIRED)
+find_package(Eigen3 REQUIRED)
 if(BUILD_TESTING)
   find_package(Boost REQUIRED COMPONENTS unit_test_framework)
   find_package(example-robot-data REQUIRED)
+endif()
+
+if(BUILD_PYTHON_INTERFACE)
+  find_package(eigenpy 2.7.10 REQUIRED)
+  string(REGEX REPLACE "-" "_" PY_NAME ${PROJECT_NAME})
+  set(${PY_NAME}_INSTALL_DIR ${PYTHON_SITELIB}/${PY_NAME})
 endif()
 
 # Main library
@@ -60,10 +76,35 @@ set(${PROJECT_NAME}_SOURCES
 )
 add_library(${PROJECT_NAME} SHARED ${${PROJECT_NAME}_SOURCES}
                                    ${${PROJECT_NAME}_HEADERS})
+target_link_libraries(${PROJECT_NAME} PUBLIC Eigen3::Eigen)
 target_include_directories(${PROJECT_NAME} PUBLIC $<INSTALL_INTERFACE:include>)
 target_link_libraries(${PROJECT_NAME} PUBLIC pinocchio::pinocchio)
+
 if(SUFFIX_SO_VERSION)
   set_target_properties(${PROJECT_NAME} PROPERTIES SOVERSION ${PROJECT_VERSION})
+endif()
+
+# Python Bindings
+# if(BUILD_PYTHON_INTERFACE)
+#   add_subdirectory(python)
+# endif()
+if(BUILD_PYTHON_INTERFACE)
+  set(${PY_NAME}_SOURCES
+    python/contact_6d.cpp
+  )
+  add_library(${PY_NAME}_py SHARED ${${PY_NAME}_SOURCES})
+  target_link_libraries(${PY_NAME}_py PUBLIC ${PROJECT_NAME} eigenpy::eigenpy)
+  # target_include_directories(${PY_NAME}_py SYSTEM PRIVATE ${PYTHON_INCLUDE_DIR})
+  set_target_properties(${PY_NAME}_py PROPERTIES
+    INSTALL_RPATH "$ORIGIN/../../.."
+    LIBRARY_OUTPUT_DIRECTORY ${PY_NAME}
+    OUTPUT_NAME ${PY_NAME}
+    PREFIX ""
+    SUFFIX ${PYTHON_EXT_SUFFIX}
+  )
+  TARGET_LINK_BOOST_PYTHON(${PY_NAME}_py PRIVATE)
+  install(TARGETS ${PY_NAME}_py DESTINATION ${${PY_NAME}_INSTALL_DIR})
+  install(DIRECTORY python/aig DESTINATION ${${PY_NAME}_INSTALL_DIR}/..)
 endif()
 
 if(BUILD_TESTING)
@@ -85,7 +126,11 @@ if(BUILD_TESTING)
 endif()
 
 # Installation
-install(
-  TARGETS ${PROJECT_NAME}
-  EXPORT ${TARGETS_EXPORT_NAME}
-  DESTINATION lib)
+if(NOT INSTALL_PYTHON_INTERFACE_ONLY)
+  install(
+    TARGETS ${PROJECT_NAME}
+    EXPORT ${TARGETS_EXPORT_NAME}
+    DESTINATION lib)
+  install(DIRECTORY include/ DESTINATION include)
+  install(FILES package.xml DESTINATION share/${PROJECT_NAME})
+endif()
