+++ python/aig/__init__.py
@@ -1,0 +1,1 @@
+from .aig import Contact6DSettings
